mark0/requirements.txt
alembic
bcrypt
blinker
certifi
charset-normalizer
click
Faker
Flask
Flask-Bcrypt
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
gunicorn
idna
itsdangerous
Jinja2
Mako
MarkupSafe
numpy
packaging
pandas
psycopg2-binary
python-dateutil
python-decouple
pytz
requests
six
SQLAlchemy
typing_extensions
tzdata
urllib3
Werkzeug
